move validation code from postuser file to seperate file
